rename some of the checktime files
