+++ libraries/chain/checktime_timer_macos.cpp
@@ -1,5 +1,5 @@
 #include <eosio/chain/checktime_timer.hpp>
-#include <eosio/chain/checktime_timer_calibrate.hpp>
+#include <eosio/chain/checktime_timer_accuracy.hpp>
 
 #include <fc/time.hpp>
 #include <fc/fwd_impl.hpp>
